# docs/swagger.yaml
basePath: /
definitions:
  core.LoginSuccess:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
  core.MessageResponse:
    properties:
      message:
        type: string
    type: object
  core.Product:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      quantity:
        type: integer
      updatedAt:
        type: string
    type: object
  core.UserLogin:
    properties:
      password:
        example: Pass@1234
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for a Product API.
  title: Swagger API
  version: "1.0"
paths:
  /product:
    get:
      consumes:
      - application/json
      description: Get details of all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.Product'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get all products
      tags:
      - product
  /product/{id}:
    get:
      consumes:
      - application/json
      description: Get details of product
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Product'
      security:
      - ApiKeyAuth: []
      summary: Get product
      tags:
      - product
  /user:
    post:
      consumes:
      - application/json
      description: Create user
      parameters:
      - description: Username/password
        in: body
        name: userLogin
        required: true
        schema:
          $ref: '#/definitions/core.UserLogin'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/core.MessageResponse'
      summary: Create user
      tags:
      - user
  /user/login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: Username/password
        in: body
        name: userLogin
        required: true
        schema:
          $ref: '#/definitions/core.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.LoginSuccess'
      summary: Login user
      tags:
      - user
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"